refactor: move config to root, logger to utils
